#9 inseridas 2 formas de bloqueio para duplicidade de pedidos: por javascript e por banco (com o campo mw_reserva.id_pedido_venda)
git-svn-id: https://portal.cc.com.br:8084/svn/projetos@309 236fd033-a1ae-4692-b96d-e3c3cfb81559

+++ Site/comprar/processarDadosCompra.php
@@ -404,7 +404,9 @@ if (isset($retorno['codigo_erro'])) {
 		setcookie('ipagareError['.$key.']', $val, $cookieExpireTime);
 	}
 
-	executeSQL($mainConnection, 'DELETE FROM MW_RESERVA WHERE ID_SESSION = ?', array(session_id()));
+	if ($retorno['codigo_erro'] == '201') {
+	    executeSQL($mainConnection, 'DELETE FROM MW_RESERVA WHERE ID_SESSION = ?', array(session_id()));
+	}
 	
 	ob_end_clean();
 	header("Location: pagamento_cancelado.php");
